imagine_cup/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic>=2.8.0
python-dotenv==1.0.0
